# docs/swagger.yaml
basePath: /
definitions:
...
  models.BankPaymentResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      paid_at:
        type: string
      payment_method:
        type: string
      return_id:
        type: string
      sale_id:
        type: string
      transaction_ref:
        type: string
      updated_at:
        type: string
    type: object
...
  models.DiscountUsageResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      discount_id:
        type: string
      id:
        type: string
      sale_id:
        type: string
      used_at:
        type: string
    type: object
...
  models.ReturnResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.ReturnItemResponse'
        type: array
      return_date:
        type: string
      sale_id:
        type: string
      status:
        type: string
      total_refund:
        type: number
      updated_at:
        type: string
    type: object
...
  models.UpdateSaleStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  utils.ErrorResponseModel:
    properties:
      error:
        example: Detailed error message
        type: string
      message:
        example: Error occurred
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: info@kisanlink.in
    name: Kisanlink Support
    url: https://github.com/Kisanlink/fpo-erp
  description: Comprehensive ERP system for agricultural cooperatives with multi-tenant
    architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kisanlink ERP API
  version: "1.0"
paths:
...
  /api/v1/bank-payments/sale/{saleID}:
    get:
      description: Retrieve all bank payments for a specific sale
      parameters:
      - description: Sale ID
        example: SALE_12345678
        in: path
        name: saleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bank payments for sale retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.BankPaymentResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Bank Payments by Sale
      tags:
      - Bank Payments
...
  /api/v1/discounts/usage/sale/{saleID}:
    get:
      description: Retrieve all discount usage for a specific sale
      parameters:
      - description: Sale ID
        example: SALE_12345678
        in: path
        name: saleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Discount usage retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DiscountUsageResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Discount Usage by Sale
      tags:
      - Discounts
...
  /api/v1/returns/sale/{saleID}:
    get:
      description: Retrieve all returns for a specific sale
      parameters:
      - description: Sale ID
        example: SALE_12345678
        in: path
        name: saleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ReturnResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Returns by Sale ID
      tags:
      - Returns
...
  /api/v1/sales:
    get:
      description: Retrieve all sales with pagination (requires authentication)
      parameters:
      - description: 'Number of records to return (default: 10)'
        example: 10
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        example: 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sales retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SaleResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get All Sales
      tags:
      - Sales
    post:
      consumes:
      - application/json
      description: Create a new sale (requires authentication)
      parameters:
      - description: Sale data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateSaleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Sale created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SaleResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create Sale
      tags:
      - Sales
...
  /api/v1/sales/{id}/returns:
    get:
      description: Retrieve all returns associated with a specific sale
      parameters:
      - description: Sale ID
        example: SALE_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns for sale retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ReturnResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Sale not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Returns for Sale
      tags:
      - Sales
  /api/v1/sales/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a specific sale
      parameters:
      - description: Sale ID
        example: SALE_12345678
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSaleStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sale status updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SaleResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Sale not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Update Sale Status
      tags:
      - Sales
  /api/v1/sales/date-range:
    get:
      description: Retrieve sales within a specific date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        example: "2024-01-01"
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        example: "2024-12-31"
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SaleResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Sales by Date Range
      tags:
      - Sales
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"